fix pdf report
