+++ src/Components/DataViewerComponents/FrequencyChart/Chart.vue
@@ -1,75 +1,41 @@
 <template>
   <div>
     <div
+      v-if="format == 'export'"
+      class="chart-title"
+      :style="`width: ${viewWidth}px;`"
+    >
+      {{ chartTitle }}
+    </div>
+
+    <div
       :is="viewType == 'matrix' ? 'matrix-chart' : 'bar-area-chart'"
       :view-type="viewType"
       :plot-id="plotId"
-      :data="data"
+      :data="frequencyData"
       :features="features"
-      :unit="unit"
-      :start-date="startDate"
-      :end-date="endDate"
       :minimum-date="applet.minimumDate"
       :view-width="viewWidth"
-      :range="range"
-      :selected-feature="selectedFeature"
-      format="view"
+      :start-date="options.startDate"
+      :end-date="options.endDate"
+      :unit="options.unit"
+      :range="options.range"
+      :selected-feature="options.selectedFeature"
+      :format="format"
       @selectFeature="selectFeature"
       @viewDetails="viewDetails"
     />
-
-    <vue-html2pdf
-      ref="html2Pdf"
-      :show-layout="false"
-      :float-layout="true"
-      :enable-download="true"
-      :preview-modal="false"
-      :paginate-elements-by-height="3000"
-      filename="report"
-      :pdf-quality="2"
-      :manual-pagination="false"
-      pdf-format="a4"
-      pdf-orientation="landscape"
-      pdf-content-width="800px"
-      :html-to-pdf-options="{
-        margin: 70,
-        enableLinks: true,
-        html2canvas: {
-          scale: 1,
-          useCORS: true,
-        },
-        jsPDF: {
-          unit: 'pt',
-          format: 'a4',
-          orientation: 'portrait',
-        },
-      }"
-    >
-      <section slot="pdf-content">
-        <Markdown
-          class="my-4"
-          :source="markdown"
-        />
-
-        <div
-          :is="viewType == 'matrix' ? 'matrix-chart' : 'bar-area-chart'"
-          :view-type="viewType"
-          :plot-id="`pdf-${plotId}`"
-          :data="data"
-          :features="features"
-          :unit="unit"
-          :start-date="startDate"
-          :end-date="endDate"
-          :minimum-date="applet.minimumDate"
-          :view-width="550"
-          :range="range"
-          :selected-feature="selectedFeature"
-          format="export"
-        />
-      </section>
-    </vue-html2pdf>
   </div>
 </template>
+
+<style scoped>
+.chart-title {
+  margin-bottom: 25px;
+  font-size: 20px;
+  font-weight: 500;
+  text-align: center;
+}
+</style>
 
 <script>
 import * as d3 from 'd3';
@@ -109,9 +75,26 @@ export default {
       type: Boolean,
       required: true,
     },
+    format: {
+      type: String,
+      required: false,
+      default: "view"
+    },
+    options: {
+      type: Object,
+      required: true
+    }
   },
   data: function() {
     const { features, responses } = this.applet.getTokenResponses();
+
+    return {
+      features,
+      responses,
+    }
+  },
+
+  beforeMount() {
     const now = new Date();
     now.setDate(now.getDate()+1)
 
@@ -119,40 +102,54 @@ export default {
     const startDate = new Date(endDate.getTime())
     startDate.setDate(startDate.getDate() - 1)
 
-    return {
-      features,
-      responses,
-      unit: 'hour',
-      range: '1 day',
+    this.$emit('setOptions', {
+      selectedFeature: null,
       startDate,
       endDate,
-      data: aggregate(features, responses, 'hour', startDate, endDate),
-      selectedFeature: null,
-      markdown: ''
-    }
+      unit: 'hour',
+      range: '1 day'
+    })
+
+    this.data = aggregate(this.features, this.responses, 'hour', startDate, endDate);
   },
+
   computed: {
     viewWidth() {
       return this.parentWidth - 20;
+    },
+
+    chartTitle() {
+      let title = '';
+
+      if (this.options.range != 'all') {
+        title += this.options.range.replace(/\s/g, '-') + ' ';
+      }
+
+      title += `${this.viewType} chart view of behavior frequency`;
+
+      return title;
+    },
+
+    frequencyData () {
+      return aggregate(this.features, this.responses, this.options.unit, this.options.startDate, this.options.endDate);
     }
   },
   methods: {
     selectFeature(feature) {
-      this.selectedFeature = feature;
+      this.$emit('setOptions', {
+        ...this.options,
+        selectedFeature: feature
+      })
     },
 
     viewDetails({ startDate, endDate, unit, range }) {
-      this.startDate = startDate;
-      this.endDate = endDate;
-      this.unit = unit;
-      this.range = range;
-      this.data = aggregate(this.features, this.responses, unit, startDate, endDate);
-    },
-
-    exportToPdf (markdown) {
-      this.markdown = markdown;
-
-      this.$refs.html2Pdf.generatePdf()
+      this.$emit('setOptions', {
+        selectedFeature: this.options.selectedFeature,
+        startDate,
+        endDate,
+        unit,
+        range
+      })
     }
   }
 }
