--- src/Components/DataViewerComponents/TokenChart/Graph.vue
@@ -1,671 +1,0 @@
-<template>
-  <div>
-    <div class="token-header">
-      <div class="header-text my-4">You have earned:</div>
-
-      <div class="total-tokens">
-        <img :src="require('@/assets/tokens.png')" width="100" />
-
-        <span class="token-number">{{ cumulative.toLocaleString() }}</span>
-        <span>total</span>
-      </div>
-
-      <div
-        v-if="currentInterval"
-        class="yesterday-tokens"
-      >
-        <img :src="require('@/assets/token.png')" width="32" />
-        <span class="token-number">{{ pastTokens.toLocaleString() }}</span>
-        <span v-if="currentInterval.unit == 'days'">yesterday</span>
-        <span v-if="currentInterval.unit == 'weeks'">past week</span>
-        <span v-if="currentInterval.unit == 'months'">past month</span>
-        <span v-if="currentInterval.unit == 'year'">last year</span>
-      </div>
-
-      <div
-        v-if="!exportFormat"
-        class="today-tokens"
-      >
-        <div class="title">Today you'll earn at least:</div>
-        <div class="token-number">
-          <img :src="require('@/assets/token.png')" width="32" /> from {{ applet.activities.length }} activities
-        </div>
-      </div>
-    </div>
-
-    <svg
-      v-if="!exportFormat"
-      viewBox="0 0 400 43"
-      width="100%"
-      :height="parentWidth * 43 / 400"
-    >
-      <mask id="pathMask">
-        <rect x="10" y="0" width="380" height="43" fill="white"></rect>
-      </mask>
-
-      <filter id="dropShadow">
-        <feGaussianBlur in="SourceGraphic" stdDeviation="1" />
-        <feOffset dx="0" dy="1" />
-        <feMerge>
-            <feMergeNode />
-            <feMergeNode in="SourceGraphic" />
-        </feMerge>
-      </filter>
-      <path
-        fill="rgb(200, 200, 200)"
-        :d="splitPath"
-        mask="url(#pathMask)"
-        filter="url(#dropShadow)"
-      />
-      <path
-        :fill="pathColor"
-        :d="splitPath"
-        mask="url(#pathMask)"
-      />
-    </svg>
-
-    <div
-      class="token-content"
-      :class="exportFormat ? 'pdf-export' : ''"
-      :style="`width: ${parentWidth}px;`"
-    >
-      <div class="range-options">
-        <div
-          v-for="(option, index) in rangeOptions"
-          :key="index"
-          class="option"
-          :class="option == range ? 'selected': ''"
-          @click="$emit('set-range', option)"
-        >
-          {{ option }}
-        </div>
-      </div>
-
-      <div
-        class="token-chart"
-      >
-        <v-card
-          class="tooltip"
-          :style="`display: ${tooltip.visible ? 'block' : 'none'}; left: ${tooltip.left}px; top: ${tooltip.top}px;`"
-        >
-          <span class="label">Cumulative:</span>
-          <span class="value">{{ tooltip.cumulative }}</span>
-        </v-card>
-
-        <svg
-          :id="`token-chart-${plotId}`"
-          :width="parentWidth"
-          :height="height + axisHeight"
-        >
-          <polygon
-            :fill="fillColor"
-            :points="cumulativePolygon"
-          />
-
-          <path
-            :d="cumulativePath"
-            :stroke="pathColor"
-            :stroke-width="strokeWidth"
-            fill="transparent"
-          />
-
-          <circle
-            :cx="cumulativePoints[0].x"
-            :cy="cumulativePoints[0].y-strokeWidth/2"
-            r="7"
-            fill="rgb(254, 221, 96)"
-          />
-
-          <line
-            :x1="padding.left"
-            :y1="exportFormat ? axisHeight + height : height"
-            :y2="exportFormat ? axisHeight + height : height"
-            :x2="width + padding.left"
-            stroke="rgb(162, 162, 162)"
-            stroke-dasharray="1 2"
-          />
-
-          <g class="tick-lines">
-            <line
-              v-for="(tick, index) in ticks"
-              :key="index"
-              :x1="tick.x"
-              :x2="tick.x"
-              :y1="exportFormat ? axisHeight - tickHeight : height"
-              :y2="exportFormat ? axisHeight : height + tickHeight"
-              stroke="rgb(162, 162, 162)"
-              stroke-width="1"
-            />
-          </g>
-
-          <g
-            v-if="range == 'Today'"
-            class="responses"
-          >
-            <template
-              v-for="(point, index) in cumulativePoints"
-            >
-              <template
-                v-if="!point.spend"
-              >
-                <circle
-                  :key="`circle-${index}`"
-                  :cx="point.x"
-                  :cy="height"
-                  r="7"
-                  :fill="pathColor"
-                />
-
-                <polygon
-                  :key="`star-${index}`"
-                  :points="getStarCoordinates(point.x, height + axisHeight - 5, 5)"
-                  fill="white"
-                />
-              </template>
-            </template>
-          </g>
-
-          <g class="tick-labels">
-            <text
-              v-for="(tick, index) in ticks"
-              :key="index"
-              :x="tick.x"
-              :y="exportFormat ? (axisHeight-tickHeight)/2 : height + (axisHeight + tickHeight) / 2"
-              text-anchor="middle"
-              font-size="15"
-              @mouseover="hoverTick(tick)"
-              @mouseout="tooltip.visible = false"
-            >
-              {{ tick.text }}
-            </text>
-          </g>
-        </svg>
-      </div>
-    </div>
-  </div>
-</template>
-
-<style lang="scss" scoped>
-.token-header {
-  width: 90%;
-  margin: auto;
-  position: relative;
-
-  .header-text {
-    font-size: 22px;
-    font-weight: 500;
-  }
-  .total-tokens, .yesterday-tokens {
-    display: flex;
-    align-items: flex-end;
-    color: rgb(252, 185, 45);
-  }
-
-  .total-tokens {
-    span {
-      padding-bottom: 10px;
-      margin-left: 5px;
-    }
-    .token-number {
-      padding-bottom: 0px;
-      font-size: 45px;
-      font-weight: bold;
-    }
-  }
-
-  .yesterday-tokens {
-    margin-left: 25px;
-    span {
-      margin-left: 5px;
-      font-weight: 500;
-      font-size: 18px;
-    }
-    .token-number {
-      font-size: 20px;
-      font-weight: bold;
-    }
-  }
-
-  .today-tokens {
-    position: absolute;
-    background-color: rgb(83, 83, 83);
-    border-radius: 10px;
-    right: 5px;
-    top: 5px;
-
-    padding: 10px;
-
-    .title {
-      color: white;
-    }
-
-    .token-number {
-      color: rgb(218, 164, 44);
-      display: flex;
-      align-items: center;
-    }
-  }
-}
-
-.range-options {
-  display: flex;
-  justify-content: space-around;
-
-  .option {
-    cursor: pointer;
-    font-weight: bold;
-
-    width: 60px;
-    padding: 10px 0px;
-    text-align: center;
-    border-radius: 5px;
-  }
-
-  .option.selected, .option:hover {
-    background-color: rgb(254, 221, 96);
-  }
-}
-
-.token-chart {
-  position: relative;
-}
-
-.tooltip {
-  z-index: 9999;
-  position: absolute;
-  max-width: 300px;
-  font-size: 14px;
-  flex-wrap: wrap;
-  padding: 15px;
-  transform: translateY(-100%) translateX(-50%);
-}
-
-.token-content {
-  display: flex;
-  flex-direction: column;
-
-  &.pdf-export {
-    flex-direction: column-reverse;
-    margin-top: 40px;
-
-    .range-options {
-      margin-bottom: 25px;
-    }
-  }
-}
-</style>
-
-
-<script>
-import * as d3 from 'd3';
-import * as moment from 'moment';
-import Applet from '../../../models/Applet';
-import Item from '../../../models/Item';
-
-/**
- * TokenChart component.
- */
-export default {
-  name: 'TokenChart',
-  /**
-   * Component properties.
-   */
-  props: {
-    plotId: {
-      type: String,
-      required: true,
-    },
-    applet: {
-      type: Object,
-      required: true,
-    },
-    hasVersionBars: {
-      type: Boolean,
-      required: true,
-    },
-    parentWidth: {
-      type: Number,
-      required: true,
-    },
-    range: {
-      type: String,
-      required: true
-    },
-    exportFormat: {
-      type: Boolean,
-      required: false,
-      default: false
-    }
-  },
-
-  data () {
-    const now = new Date();
-    now.setDate(now.getDate()+1)
-
-    const endDate = new Date(moment(now).format('YYYY-MM-DD'));
-
-    return {
-      endDate,
-      rangeOptions: [
-        'Today',
-        '1w',
-        '2w',
-        '1m',
-        '3m',
-        '1y',
-        'All'
-      ],
-      height: 250,
-      axisHeight: 40,
-      tickHeight: 6,
-      padding: { left: 40, right: 40 },
-      strokeWidth: 4,
-      fillColor: 'rgb(254, 239, 180)',
-      pathColor: 'rgb(252, 185, 45)',
-      tooltip: {
-        visible: false,
-        left: 0,
-        top: 0,
-        cumulative: 0
-      }
-    }
-  },
-
-  computed: {
-    width () {
-      return this.parentWidth - this.padding.left - this.padding.right;
-    },
-
-    availableIntervals () {
-      return [
-        { amount: 1, unit: 'days', range: 'Today' },
-        { amount: 1, unit: 'weeks', range: '1w' },
-        { amount: 2, unit: 'weeks', range: '2w' },
-        { amount: 1, unit: 'months', range: '1m' },
-        { amount: 3, unit: 'months', range: '3m' },
-        { amount: 1, unit: 'year', range: '1y' }
-      ]
-    },
-
-    currentInterval () {
-      return this.availableIntervals.find(iv => iv.range == this.range)
-    },
-
-    startDate () {
-      const endTime = moment.utc(this.endDate);
-
-      switch (this.range) {
-        case 'Today':
-          return endTime.subtract(1, 'days').toDate()
-        case '1w':
-          return endTime.subtract(1, 'weeks').toDate()
-        case '2w':
-          return endTime.subtract(2, 'weeks').toDate()
-        case '1m':
-          return endTime.subtract(1, 'months').toDate()
-        case '3m':
-          return endTime.subtract(3, 'months').toDate()
-        case '1y':
-          return endTime.subtract(1, 'years').toDate()
-      }
-
-      for (const interval of this.availableIntervals) {
-        if (endTime.diff(this.applet.minimumDate, interval.unit) < interval.amount) {
-          return endTime.subtract(interval.amount, interval.unit).toDate()
-        }
-      }
-
-      return new Date(`${endTime.format('YYYY-MM')}-01`);
-    },
-
-    ticks () {
-      const endTime = moment(this.endDate)
-      let range = this.range;
-
-      if (range == 'All') {
-        range = '1y'
-
-        for (const interval of this.availableIntervals) {
-          if (endTime.diff(this.startDate, interval.unit) <= interval.amount ) {
-            range = interval.range;
-            break;
-          }
-        }
-      }
-
-      let unit = 'months', ticks = [];
-
-      if (range == 'Today') {
-        const date = moment.utc(this.startDate).format('YYYY-MM-DD')
-
-        ticks = [
-          {
-            time: moment.utc(this.startDate),
-            text: moment.utc(this.startDate).format('ddd M/DD'),
-          },
-          { time: moment.utc(`${date} 09:00`, 'YYYY-MM-DD HH:mm'), text: 'Morning' },
-          { time: moment.utc(`${date} 12:00`, 'YYYY-MM-DD HH:mm'), text: 'Noon' },
-          { time: moment.utc(`${date} 18:00`, 'YYYY-MM-DD HH:mm'), text: 'Evening' },
-          {
-            time: moment.utc(this.endDate),
-            text: moment.utc(this.endDate).format('ddd M/DD'),
-          },
-        ]
-      } else {
-        switch (range) {
-          case '1w': case '2w': case '1m':  // day view
-            unit = 'days';
-            break;
-          case '3m':  // week view
-            unit = 'weeks';
-            break;
-          case '1y': // month view
-            unit = 'months';
-            break;
-        }
-
-        let tick = moment.utc(this.startDate)
-
-        if (range == '1y') {
-          tick = moment.utc(tick.format('YYYY-MM') + '01', 'YYYY-MM-DD')
-        }
-
-        while (tick.unix() <= endTime.unix()) {
-          if (tick.unix()*1000 >= this.startDate.getTime()) {
-            ticks.push({
-              time: moment.utc(tick)
-            })
-          }
-
-          tick = tick.add(1, unit)
-        }
-
-        ticks = ticks.map((tick, index) => {
-          let text = tick.time.format(!index || index == ticks.length-1 ? 'MMM DD' : 'MMM');
-          if (unit == 'days' || unit == 'weeks') {
-            text = tick.time.format(!index || index == ticks.length-1 ? 'MMM D' : 'D');
-          }
-          return { ...tick, text }
-        })
-      }
-
-      return ticks.map(tick => ({
-        ...tick,
-        x: this.getX(tick.time.unix() * 1000)
-      }))
-    },
-
-    pastTokens () {
-      if (!this.currentInterval) {
-        return 0;
-      }
-
-      const unit = this.currentInterval ? this.currentInterval.unit : 'days';
-      const range = [
-        moment.utc(this.endDate).subtract(1, unit).toDate(),
-        moment.utc(this.endDate).subtract(2, unit).toDate()
-      ]
-
-      let tokens = 0;
-
-      for (const change of this.applet.token.changes) {
-        if (change.time < range[0].getTime() && change.time > range[1].getTime()) {
-          tokens += change.value;
-        }
-      }
-
-      return tokens;
-    },
-
-    todayTokens () {
-      const today = moment.utc(this.endDate).subtract(1, 'days').toDate();
-      let tokens = 0;
-
-      for (const change of this.applet.token.changes) {
-        if (change.time > today) {
-          tokens += change.value;
-        }
-      }
-
-      return tokens;
-    },
-
-    cumulative () {
-      return this.applet.token.cumulative
-    },
-
-    history () {
-      const startTime = this.startDate.getTime();
-      const endTime = this.endDate.getTime();
-
-      return this.applet.token.changes
-              .filter(({ time }) => time >= startTime && time <= endTime)
-    },
-
-    yUnit () {
-      let maxCumulative = this.cumulative, current = this.cumulative;
-
-      for (const change of this.applet.token.changes) {
-        current -= change.value;
-
-        if (maxCumulative < current) {
-          maxCumulative = current;
-        }
-      }
-
-      return Math.min(25, (this.height - 20) / maxCumulative);
-    },
-
-    cumulativePoints () {
-      const startTime = this.startDate.getTime();
-      const endTime = Math.min(this.endDate.getTime(), new Date().getTime());
-
-      const changes = [
-        { time: endTime, value: 0, spend: true },
-        ...this.history,
-        { time: startTime, value: 0, spend: true },
-      ];
-
-      const points = []
-      let cumulative = this.cumulative;
-
-      for (let i = 0; i < changes.length; i++) {
-        if (changes[i].value) {
-          points.push({
-            x: this.getX(changes[i].time),
-            y: this.height - this.yUnit * cumulative + (this.exportFormat ? this.axisHeight : 0),
-            spend: changes[i].spend
-          })
-        }
-
-        cumulative -= changes[i].value;
-
-        points.push({
-          x: this.getX(changes[i].time),
-          y: this.height - this.yUnit * cumulative + (this.exportFormat ? this.axisHeight : 0),
-          spend: changes[i].spend
-        })
-      }
-
-      return points;
-    },
-
-    cumulativePath () {
-      const points = this.cumulativePoints;
-
-      return points.map((point, index) => `${!index ? 'M' : 'L'} ${point.x} ${point.y - this.strokeWidth/2}`).join(' ')
-    },
-
-    cumulativePolygon () {
-      const points = [...this.cumulativePoints];
-
-      const endTime = Math.min(this.endDate.getTime(), new Date().getTime());
-
-      points.push({
-        x: this.getX(this.startDate.getTime()),
-        y: this.height + (this.exportFormat ? this.axisHeight : 0)
-      })
-
-      points.push({
-        x: this.getX(endTime),
-        y: this.height + (this.exportFormat ? this.axisHeight : 0)
-      })
-
-      return points.map(({ x, y }) => `${x}, ${y}`).join(' ')
-    },
-
-    splitPath () {
-      return `M 1.00,42.00 C 1.08,39.61 0.95,37.38 2.60,35.42 4.34,33.34 8.44,32.14 11.00,31.35 11.00,31.35 32.00,26.21 32.00,26.21 65.58,19.06 99.78,17.60 134.00,18.00
-        134.00,18.00 149.00,19.00 149.00,19.00 149.00,19.00 164.00,19.91 164.00,19.91 164.00,19.91 211.00,24.17 211.00,24.17
-        211.00,24.17 227.00,26.00 227.00,26.00 227.00,26.00 236.00,26.00 236.00,26.00 236.00,26.00 248.00,27.00 248.00,27.00 248.00,27.00 284.00,27.00 284.00,27.00
-        284.00,27.00 296.00,26.09 296.00,26.09 317.21,24.64 337.33,21.62 358.00,16.47 358.00,16.47 369.00,13.96 369.00,13.96 379.55,10.83 386.23,7.21 395.97,3.00
-        396.36,4.72 396.92,5.87 395.97,7.61 394.77,9.67 391.13,10.96 389.00,11.77 389.00,11.77 372.00,17.66 372.00,17.66 359.77,21.57 329.44,27.85 317.00,28.00
-        317.00,28.00 285.00,30.00 285.00,30.00 285.00,30.00 246.00,30.00 246.00,30.00 246.00,30.00 225.00,29.00 225.00,29.00 225.00,29.00 212.00,28.00 212.00,28.00
-        212.00,28.00 159.00,25.00 159.00,25.00 159.00,25.00 142.00,24.00 142.00,24.00 142.00,24.00 112.00,24.00 112.00,24.00 112.00,24.00 93.00,25.00 93.00,25.00 81.25,25.14 47.41,29.04 36.00,31.89
-        36.00,31.89 16.00,37.70 16.00,37.70 16.00,37.70 7.00,41.21 7.00,41.21 4.80,41.84 3.24,41.89 1.00,42.00 Z`
-    }
-  },
-
-  methods: {
-    hoverTick (tick) {
-      const index = this.ticks.indexOf(tick)
-      const startTime = tick.time.unix() * 1000;
-      const endTime = startTime + 86400 * 1000;
-
-      let tokens = 0;
-
-      for (const change of this.applet.token.changes) {
-        if (change.time < endTime && change.time > startTime) {
-          tokens += change.value;
-        }
-      }
-
-      this.tooltip = {
-        visible: true,
-        cumulative: tokens,
-        left: tick.x,
-        top: this.height - 5
-      }
-    },
-
-    getStarCoordinates (cx, cy, r) {
-      const diff = 144 / 180 * Math.PI;
-      const points = [];
-
-      for (let i = 0; i < 5; i++) {
-        const angle = Math.PI/2 - diff * i;
-        points.push({ x: Math.cos(angle) * r + cx, y: Math.sin(angle) * r + cy })
-      }
-
-      return points.map(({x, y}) => `${x}, ${y}`).join(' ')
-    },
-
-    getX (time) {
-      const startTime = this.startDate.getTime();
-      const endTime = this.endDate.getTime();
-
-      return this.width * (time - startTime) / (endTime - startTime) + this.padding.left;
-    }
-  },
-};
-</script>
